# Fix CLP(FD) multiplication leaving a lot of residual variables