@JanWielemaker 2021-08-04
@triska, what do you think?    I think there is some room for optimizing with_local_attributes/4.   First of all, the complex goals are passed as Prolog terms.   Turning the body into a predicate should make this a little more efficient.   See also the `compile_meta_arguments` flag.

Another "solution" might be to make these changes dependent on whether or not we are in scope of call_residue_vars/2.   There is currently no way to find out, but that can be added easily as there is a global flag at the C level that affects attribute handling and garbage collection when call_residue_vars/2 is active.