@triska 2021-08-04
This is definitely a good change, thank you a lot for this improvement!

The overarching goal (for the far future) is to wrap all toplevel queries implicitly in `call_residue_vars/2`, so that everything gets truthfully reported. Every change that makes this eventual development easier is a good one!

The "counterargument" so far is that using `call_residue_vars/2` implicitly is too slow. And it's easy to see why that is regarded as a valid argument: Because who cares about the truth of the answer, as long as we get it fast, right?

One minor comment about terminology: I use "FIXED" in commits that correct declarative mistakes, such as an invalid pruning of values. This concrete change is more of an enhancement than a correction, because retaining and reporting these residual goals does not change the truth of the answer. It is definitely good for performance to forget these constraints as they are no longer needed after the propagation.