@dgelessus 2021-08-05
I played around with this some more. For some reason, combining the two `with_local_attributes` blocks improves the performance significantly - now it's only 1.5x as slow as before, rather than 2.5x. Still not great, but much better than before.

I also tried declaring `with_local_attributes` as a meta-predicate - that alone caused a very small performance decrease. But if I then do `set_prolog_flag(compile_meta_arguments, control)` before loading `library(clpfd)`, it becomes a bit faster than before.

For comparison, here are the rough times I get for the `time((between(1, 1000, _), 10 #= Y * Z, fail))` test case:

| Version | Times |
|---|---|
| No fix | 7,940,001 inferences, 1.120 CPU in 1.128 seconds (99% CPU, 7090901 Lips) |
| Original fix | 16,452,001 inferences, 2.879 CPU in 2.942 seconds (98% CPU, 5713631 Lips) |
| With combined with_local_attributes calls | 9,232,001 inferences, 1.534 CPU in 1.571 seconds (98% CPU, 6018540 Lips) |
| With `meta_predicate` | 9,280,001 inferences, 1.635 CPU in 1.669 seconds (98% CPU, 5676283 Lips) |
| With `compile_meta_arguments` | 9,232,001 inferences, 1.455 CPU in 1.491 seconds (98% CPU, 6343091 Lips) |

> The "counterargument" so far is that using `call_residue_vars/2` implicitly is too slow.

Well, I only discovered this problem because of an application that frequently uses call_residue_vars/2 to ensure that there are no leftover residual constraints, and it seems to be performing fine 🙂 I was actually worrying that "no residual goals ever" might be too extreme of a check, with how easy it is for internal constraints like these to "leak" out via call_residue_vars/2.

> This concrete change is more of an enhancement than a correction, because retaining and reporting these residual goals does not change the truth of the answer.

You are right that the extra constraints don't change the correctness of the answer. (Assuming the extra constraints can actually be solved - I never checked...) But my commit messages here aren't final anyway - once we're done figuring out a good fix I'll probably squash everything and rewrite the commit message.