@JanWielemaker 2021-08-05
As is, with_local_attributes solves, and then throws the result as an answer.  This causes a rollback.  throw/catch aren't that fast though.   An alternative approach might be to use

```
     State = s(0),
     (    <guarded code>,
          nb_setarg(1, State, Result),
          fail
     ;    arg(1, State, Result),
          Result \== 0
     )
```

If you define this as goal expansion you also get nice compiled version for the _guarded code_.   I __think__ the above is logically equivalent to the current code, provided the `fail` doesn't cause any backtracking in to _guarded code_.  A simple `(guarded code -> true)`  can resolve that.   Throw/1 and nb_setarg/3 use the same mechanism to guard the term against backtracking.

This is a little less portable, although many systems can do a similar trick.   A small advantage is that a debug session that traps exceptions won't be interrupted.

I'll keep tracking this issue and wait for a signal that it is ready for merging.    Thanks!