@dgelessus 2021-08-10
Well, changing `with_local_attributes/4` to use `nb_setarg` seems to improve performance a little bit, but not very much:

```prolog
% With throw/catch
?- time((between(1, 1000, _), 10 #= Y * Z, fail)). % (5 times)
% 9,232,027 inferences, 1.447 CPU in 1.473 seconds (98% CPU, 6381955 Lips)
% 9,232,001 inferences, 1.487 CPU in 1.517 seconds (98% CPU, 6208595 Lips)
% 9,232,001 inferences, 1.455 CPU in 1.483 seconds (98% CPU, 6343488 Lips)
% 9,232,001 inferences, 1.487 CPU in 1.518 seconds (98% CPU, 6209100 Lips)
% 9,232,001 inferences, 1.454 CPU in 1.484 seconds (98% CPU, 6349600 Lips)

% With nb_setarg
?- time((between(1, 1000, _), 10 #= Y * Z, fail)). % (5 times)
% 9,192,028 inferences, 1.348 CPU in 1.361 seconds (99% CPU, 6818825 Lips)
% 9,192,001 inferences, 1.440 CPU in 1.471 seconds (98% CPU, 6381349 Lips)
% 9,192,001 inferences, 1.390 CPU in 1.406 seconds (99% CPU, 6611704 Lips)
% 9,192,001 inferences, 1.451 CPU in 1.478 seconds (98% CPU, 6334052 Lips)
% 9,192,001 inferences, 1.438 CPU in 1.459 seconds (99% CPU, 6391937 Lips)
```

I also tried it with ProB's unit tests and the performance difference isn't really noticeable. With the `nb_setarg` solution, the unit tests also give me errors about remaining residual goals which I can't figure out right now (this didn't happen with the previous `throw`/`catch`-based solution). Probably I'm just using `nb_setarg` wrong - I'm not really familiar with its exact behavior and pitfalls in all cases...

In any case, the `throw`/`catch`-based solution is fast enough IMO, so feel free to merge this.